rename simulations to simulator
