+++ api/src/opentrons/api/session.py
@@ -22,7 +22,7 @@ from opentrons.commands import types as command_types, introspection
 from opentrons.protocols.api_support.types import APIVersion
 from opentrons.protocols.context.protocol_api.protocol_context import \
     ProtocolContextImplementation
-from opentrons.protocols.context.simulators.protocol_context import \
+from opentrons.protocols.context.simulator.protocol_context import \
     ProtocolContextSimulation
 from opentrons.protocols.parse import parse
 from opentrons.protocols.types import Protocol
