+++ api/src/opentrons/protocols/context/simulator/protocol_context.py
@@ -3,7 +3,7 @@ from opentrons.protocols.context.interfaces.instrument_context import \
     InstrumentContextInterface
 from opentrons.protocols.context.protocol_api.protocol_context import \
     ProtocolContextImplementation
-from opentrons.protocols.context.simulators.instrument_context import \
+from opentrons.protocols.context.simulator.instrument_context import \
     InstrumentContextSimulation
