--- api/src/opentrons/protocols/context/simulators/__init__.py
@@ -1,1 +1,0 @@
-"""Context simulations."""
