+++ api/tests/opentrons/protocol_api/test_context.py
@@ -8,7 +8,7 @@ import opentrons.protocols.api_support as papi_support
 import opentrons.protocols.geometry as papi_geometry
 from opentrons.protocols.context.protocol_api.protocol_context import \
     ProtocolContextImplementation
-from opentrons.protocols.context.simulators.protocol_context import \
+from opentrons.protocols.context.simulator.protocol_context import \
     ProtocolContextSimulation
 from opentrons_shared_data import load_shared_data
 from opentrons.types import Mount, Point, Location, TransferTipPolicy
